Compiling for Edison generated duplicated symbols
The .cpp was included instead of the header generated duplicated
symbols when compiling for Edison.

Also adding extra information about how to compile for Edison.

Added clarification about the folder to set mraa.h
The Intel license headers are not the same as the Samsung ones,
so the --no_check_tidy is required if added into the iotjs project folder.

Original Edison libmraa.so requires libft4222.so symbols to link.

Signed-off-by: Martinez Rodriguez, Sergio <sergio.martinez.rodriguez@intel.com>