Enable TCP for NuttX
* On Nuttx Side, EOF doesn't come with the event when polled.
In order to workaround this issue, when called `Socket.prototype.end`,
I marked the last chunk with magic string, `\\e\\n\\d` to check if it is the last
buffer.
* For this implementation, `writeAfterEnd` should be checked with only flag `state.ended`,
because when ReadableStream is ended, it calls `onSocketEnd`.
In the default mode, `state.allowHalfOpen` is false, it calls
`destroySoon`, in the end, automatically calls `Socket.prototype.end`.
Therfore, for example, it is possible when socket receiving `end` event,
`Socket.prototype.end` was already called so that it might raise an error
`write after end`.

IoT.js-DCO-1.0-Signed-off-by: wonyong.kim wonyong.kim@samsung.com