Handle exceptions in process._onUncaughtException
Uncaught exceptions trigger the execution of
process._onUncaughtException (implemented in JS). However, if that
function throws, that causes an assertion failure (in C). This
patch fixes the issue by allowing exceptions in the handler and
exiting (but not aborting) as the last resort.

The patch also fixes a type on the name of _onUncaughtException.

Fixes #648.

IoT.js-DCO-1.0-Signed-off-by: Akos Kiss akiss@inf.u-szeged.hu