Detect supported modules when running the tests
The patch will allow the test framework to automatically
detect which modules are built-in the IoT.js. Based on this
information it can check if the input test file contains
one of the supported module name and if not the test can
be skipped.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com