[Code Coverage] Covering buffer + removing unneccessary functions
Buffer.js got a 100% coverage.
In addition there was an unnecessary if statement which checked whether the object calling it is a Buffer or not. This is not needed, since the function itself can only be called from a Buffer object, therefore this test would never run.
In iotjs_module_buffer.c IOTJS_ASSERT made the script crash, thus got commented.

IoT.js-DCO-1.0-Signed-off-by: Daniel Balla dballa@inf.u-szeged.hu