Do not create empty object for native js sources (#795)
For each js native source an object was created and the
native handle was set for it. However the native handle
was never queried.

For now removing the object creation and just setting the
value to a boolean true. It would be possible to use a simple
list also but for that the module loading needs a bit of work
as currently it relies on the fact that the native_sources are
an object witk names as keys and something true as values.

Reduces base peak memory usage a bit.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com