Fix Writable Stream's options.highWaterMark
The previous code caused the program to stop if options was not defined while creating a Writable Stream.
This reverts to the default highWaterMark if options is undefined.
The options argument for Writable Constructor should ideally be an optional argument, since the only property it must supply is highWaterMark, which we have a default value for.

IoT.js-DCO-1.0-Signed-off-by: Akhil akhil.kedia@samsung.com