Rename test_promise.js to test_iotjs_promise.js (#869)
* Rename test_promise.js to test_iotjs_promise.js

As per #743 the test_promise.js name for the test
would expect that the iotjs would be built with
promise iotjs(!) module. However the promise is
not an iotjs feature but it is provided by the
underlying JS engine JerryScript. Thus the test
runner incorrectly returns the reason for skipping
the test, stating that is not supported by the
iotjs build but that is the incorrect message.

By renaming the testcase to test_iotjs_promise.js
the correct skip reason is displayed and will
make it easier in the future to enable the test case.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com

* Extend test case writing with naming rule

Add test case naming rule to the Getting Started
document.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com