Move process argument initialization to process internals
Main points of the change:
* Removed process.argv and process._initArgv from iotjs.js.
* Reworked InitArgv method to create the process.argv array.

The main motivation was to add object type check for the InitArgv
method but by moving the argv initialization into the C side
we can avoid the check and reduce the number of properties
for the process module.

Also this should reduce the memory usage a bit. As one
function property and one magic string was removed.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com