Use an own http server for testing (#1103)
This fixes the test failure or delay upon Travis.

The old test assumes that a web server runs on the localhost with default port. So, the requested socket could be hangged up when the web server is busy. That could cause delays or timeout especially on Travis.

IoT.js-DCO-1.0-Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com