Add auto-release feature
This suggests auto-release feature for `jerry_value_t`, `iotjs_jargs_t`,
and `iotjs_string_t`. Although jerryscript has the same for `jerry_value_t`,
it seems better for us to have our own one. Because the jerry feature is
as small pieces of code as we can easily handle. Also, we need to add
more types into the implementation.

This is only applied to the `buffer` module to listen to opinions.

IoT.js-DCO-1.0-Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com