Fix portable build warning in magic strings
Usage of 'defined' in macro is not portable as reported
by the GCC compiler.

Simplifying the module checks to resolve the warning. This
is done by removing the 'defined' part. This works
because all ENABLE_MODULE_<module_name> defines are
set to at least a 0.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com