Properly terminate the engine at the first uncaughtException
There's a special case where if an uncaughtException is thrown, and later a callback function is evaluated it would throw another one.
The correct behaviour for this would be terminating the engine at the first one, and throwing the error.
This patch fixes #1349

IoT.js-DCO-1.0-Signed-off-by: Daniel Balla dballa@inf.u-szeged.hu