Add calling js functions based on string parameters
As written in iotjs_tizen_service_app.c, iot.js app for tizen is a
hybrid one, which combines features of both native and js app.
This commit provides a way to call  js function bound with tizen
module in the app model.

This also includes a patch for the bug in getResPath.

IoT.js-DCO-1.0-Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com