Hardening security of iotjs_bufferwrap_create_buffer
Currently this function calls the buffer constructor, which
can be overwritten by users. The successful construction
is only checked by an assert, and the return value of this
function is never checked.

Furthermore the memory consumption of buffers is reduced by
allocating data space after the buffer, so the buffer pointer
is removed.

IoT.js-DCO-1.0-Signed-off-by: Zoltan Herczeg zherczeg.u-szeged@partner.samsung.com