Pass the extra JerryScript build parameters for the snapshot tool build
When using `--jerry-cmake=..` argument for the build script
the arguments were not passed for the snapshot tool.
This could result in an generated incompatible snapshot for the
iotjs binary.

To fix this all `jerry-cmake` params should be passed for the
snapshot tool also.

Additonally make sure the system allocator is disabled in
case of the snapshot tool.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com