Add options argument for http createServer
Improve compatibilty with node.js by adding the `options`
argument for the createServer method in the http modules.

In addition allow "subclassing" the http.Incoming and httpServerResponse
objects. With this it is possible to augment the created request and
response objects passed for the `requestListener` method.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com