[WIP] Add iotjs_context macro and external_context build option
The external_context build option has two status ON and OFF (default is OFF). If the external_context is
on then jerry_external_context build option is also on, and in IoTjs we are able to multithread.
IOTJS_CONTEXT macro has two behaviors that depends on the external_context build option status.
Also if the external_context is on then the user has to define an iotjs_create_context and
iotjs_port_get_current_context methods.

IoT.js-DCO-1.0-Signed-off-by: Bence Gabor Kis kisbg@inf.u-szeged.hu