Fix an issue with writable streams on nuttx platform
There's an issue on nuttx systems causing a Buffer to be converted to a string everytime a `Writable stream` ends. This is obviously not a good practice, especially when working with binary buffers.
This patch fixes the issue by differentiating Buffers from strings.

IoT.js-DCO-1.0-Signed-off-by: Daniel Balla dballa@inf.u-szeged.hu