Removed an unnecessary 'do { } while(0);' from HTTP parser module. (#1839)
Usually 'do { } while(0)' without a semicolon is used in a macro when we want
to force the semicolon for the certain macro. The compiler does fails if the
developer forgot the semicolon after the macro call. Using this technique with
a semicolon after the 'while' has no sense ('...while(0);').

IoT.js-DCO-1.0-Signed-off-by: László Langó llango.u-szeged@partner.samsung.com