Simplify the TLS checks in WebSocket
The TLS checks in the WebSocket module involved requesting
the TLS module's keys as an array. This is uses extra memory,
a simple true/false check on the TLS object is enough to validate
if the 'require' call loaded the module.

IoT.js-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com