[WIP] Possible fix for closing a frozen handle
So #1907 is a pretty complex issue.
A bound udp socket gets frozen, and on that frozen object a close() is called.
Due to the freeze the object's properties can not be changed, including the native handle, namely ._handle

This fix is not nice IMHO, but solves the above issue.
I'm open to any suggestions on this PR.

IoT.js-DCO-1.0-Signed-off-by: Daniel Balla dballa@inf.u-szeged.hu